requirements.txt
# Core Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# Graph Processing
networkx>=3.2
matplotlib>=3.8.0

# Demo Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Data Processing
pandas>=2.1.0
python-dateutil>=2.8.0

# Email Parsing
email-validator>=2.1.0
beautifulsoup4>=4.12.0

# LLM & API
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.8.0
